🚚 Moved thread & presence events to pincer/events
